[Bug]- partnership page on school card open and close image opening
