+++ src/assets/program.js
@@ -1,8 +1,37 @@
 import { one, two, three, four, schoolLogo, studentImage } from ".";
 
+export const galleryImages = [
+  {
+    img: 'https://images.pexels.com/photos/1779487/pexels-photo-1779487.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1',
+    className: ' row-span-1 col-span-1 bg-white hover:opacity-[.4]',
+    id: 1,
+  },
+  {
+    img: "https://images.pexels.com/photos/3861458/pexels-photo-3861458.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    className: 'col-span-1 bg-white hover:opacity-[.4]',
+    id: 2,
+  },
+  {
+    img: "https://images.pexels.com/photos/546819/pexels-photo-546819.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    className: 'row-span-2 col-span-2 bg-white hover:opacity-[.4]',
+    id: 3,
+  },
+  {
+    img: "https://images.pexels.com/photos/1194713/pexels-photo-1194713.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    className: 'col-span-1 bg-white hover:opacity-[.4]',
+    id: 4,
+  },
+  {
+    img: "https://images.pexels.com/photos/39284/macbook-apple-imac-computer-39284.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1",
+    className: 'col-span-1 bg-white hover:opacity-[.4]',
+    id: 5,
+  },
+]
+
 export const schoolDetails = [
   {
     id: 1,
+    isOpen: false,
     bgImage: studentImage,
     logo: schoolLogo,
     number: one,
@@ -11,13 +40,14 @@ export const schoolDetails = [
     institute: "School",
     location: "Kuje, Abuja",
     prefrence: {
-            title1: "Campaign",
-            title2: "CCTV Installation",
-            title3: "Seminar",
-        },
+      title1: "Campaign",
+      title2: "CCTV Installation",
+      title3: "Seminar",
+    },
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 2,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -34,6 +64,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 3,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -50,6 +81,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 4,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -66,6 +98,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 5,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -82,6 +115,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 6,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -98,6 +132,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 7,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -114,6 +149,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 8,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -130,6 +166,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 9,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -146,6 +183,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 10,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -162,6 +200,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 11,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -178,6 +217,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 12,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -194,6 +234,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 13,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -210,6 +251,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 14,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -226,6 +268,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 15,
     bgImage: studentImage,
     logo: schoolLogo,
@@ -242,6 +285,7 @@ export const schoolDetails = [
     desc: 'The Good Shepherd Group is set to go into the next decade as a model school in Nigeria with branches in all states of the federation. In realization of this noble dream, the Ogun State campus opened up for big time academic activities in September, for the 2006/2007 session.'
   },
   {
+    isOpen: false,
     id: 16,
     bgImage: studentImage,
     logo: schoolLogo,
